# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: CafeOre POS API
  version: 1.0.0
  description: 珈琲・俺POSシステムのAPI仕様
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
...
  /api/caos/cups:
    put:
      summary: CaOS が決めたことを注文のカップに書く
      description: |
        CaOS（ドリップ管制）の盤面は、注文のカップ（OrderResponse の cups）の列（dripper・dripper_position・drip_id・brew_started_at・brew_finished_at）で持つ。
        割当・ドリッパーの移動・順番の入れ替え・未割当に戻す・統合は、どれもこの PUT でカップの列を書く。
        writes の全部を 1 つのトランザクションで書く（どれか 1 つでも通らなければ何も書かない）。
        書いたカップの注文は /api/ws/orders の {"type":"order"} で全部の画面に配り、ほかのインスタンスへは DB の通知（orders_changed）で知らせる。

        抽出の時刻は画面から送らない。空いているドリッパーに置いてそのまま始めるときは after の start_brew を true にし、サーバーが今の時刻を brew_started_at に入れる
        （iPad の時計がずれていても、残り時間がサーバーの時刻でそろう）。終えるのは「次へ」だけ。

        確かめること：
        - before が今の値と違う（ほかの端末が先に書いた・注文の編集で消えた・「次へ」で始まった）なら 409（楽観ロック）。画面は届いた注文で盤面を組み立て直す
        - 抽出中・終わりのカップ（brew_started_at のあるカップ）は書けない
        - 今日（日本時間）の注文のカップだけ書ける
        - 抽出が要らない種類（item_types.needs_brew が false）のカップは、ドリッパーにもカードにも入れられない
        - 指名の番号のあるカップ（明細の dripper が 1〜6）は、その番号のドリッパーにしか置けない
        - 限定のカップ（item_types.senior_only）は、今日の担当者が上級生のドリッパー（GET /api/caos/lanes の senior）にしか置けない。
          別のドリッパーへ置くとき（未割当から置く・ドリッパーを移す）だけ確かめ、同じドリッパーの中の順番の入れ替えは確かめない
          （担当者を上級生でない人に替えても、待っていた限定のカードはそのドリッパーに残るので）。
          指名のドリッパーの担当者が上級生でない限定のカップは、どこにも置けない
        - 1 つのドリッパーで同時に抽出中のカードは 1 枚。1 枚のカードは最大 2 杯。同じ drip_id のカップは同じ値（書かないカップも含めて）
        - 緊急のカップ（emergency_at のあるカップ）では、before・after の drip_id は入れ直しのカード（emergency_drip_id）を指す（最初に淹れた drip_id は書き換えない）。
          緊急のカップとほかのカップを 1 枚のカードにはできない
      operationId: writeCaosCups
      tags:
        - caos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaosCupsWriteRequest'
      responses:
        '204':
          description: 書いた
        '400':
          description: 形の違うリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: before が今の値と違う（何も書かない）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 決まりに合わない（何も書かない）。error を画面にそのまま出す
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/caos/drippers/{dripper}/next:
    post:
      summary: CaOS の「次へ」
      description: |
        そのドリッパーの抽出中のカードを終え（brew_finished_at）、そのカードのカップだけを準備完了にして（POS のカップの準備完了と同じく、注文の状態はカップから決め直す）、
        同じドリッパーの待機の先頭（dripper_position・注文番号の順）を抽出中にする（brew_started_at）。時刻はサーバーの今。
        抽出中が無ければ（マスターで準備完了にして終わった、など）待機の先頭を始めるだけ。
        drip_id には画面が抽出中と見ているカードを送る。今の抽出中と違えば 409（二度押しや、ほかの端末と同時に押したときに次のカードまで終わらせない）。
      operationId: advanceCaosDripper
      tags:
        - caos
      parameters:
        - name: dripper
          in: path
          required: true
          description: ドリッパーの番号（1〜6）
          schema:
            type: integer
            minimum: 1
            maximum: 6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaosNextRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaosNextResult'
        '400':
          description: 形の違うリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 画面の見ている抽出中が今と違う・抽出中も待機も無い（何も変えない）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CaosCupState:
      type: object
      description: カップの今の CaOS の値（OrderCupResponse の同じ名前の列をそのまま。緊急のカップの drip_id は emergency_drip_id）。全部 null なら未割当
      required: [dripper, dripper_position, drip_id, brew_started_at, brew_finished_at]
      properties:
        dripper:
          type: integer
          nullable: true
          minimum: 1
          maximum: 6
        dripper_position:
          type: number
          format: double
          nullable: true
        drip_id:
          type: string
          format: uuid
          nullable: true
        brew_started_at:
          type: string
          format: date-time
          nullable: true
        brew_finished_at:
          type: string
          format: date-time
          nullable: true
    CaosCupAfter:
      type: object
      description: |
        CaOS がカップに書く値。抽出の時刻は送らない（開始・終了の時刻はサーバーの今で付ける）。
        start_brew が true なら抽出を始める（brew_started_at にサーバーの今を入れる）。false なら待機・未割当（brew_started_at・brew_finished_at は null）。
        dripper・dripper_position・drip_id が全部 null で start_brew が false なら未割当
      required: [dripper, dripper_position, drip_id, start_brew]
      properties:
        dripper:
          type: integer
          nullable: true
          minimum: 1
          maximum: 6
        dripper_position:
          type: number
          format: double
          nullable: true
        drip_id:
          type: string
          format: uuid
          nullable: true
        start_brew:
          type: boolean
          description: 抽出を始める（空いているドリッパーに置いてそのまま始める）。時刻はサーバーの今
    CaosCupsWrite:
      type: object
      description: |
        カップの組を before から after にする。before はカップの今の値（注文の応答の値をそのまま送り返す。時刻はミリ秒までで比べる）。
        抽出中・終わりのカップ（brew_started_at のあるカップ）は書けない（終えるのは「次へ」）
      required: [cup_ids, before, after]
      properties:
        cup_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        before:
          $ref: '#/components/schemas/CaosCupState'
        after:
          $ref: '#/components/schemas/CaosCupAfter'
    CaosCupsWriteRequest:
      type: object
      required: [writes]
      properties:
        writes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CaosCupsWrite'
    CaosNextRequest:
      type: object
      required: [drip_id]
      properties:
        drip_id:
          type: string
          format: uuid
          nullable: true
          description: 画面が抽出中と見ているカードの drip_id。抽出中が無いと見ているなら null
    CaosNextResult:
      type: object
      required: [finished_drip_id, started_drip_id]
      properties:
        finished_drip_id:
          type: string
          format: uuid
          nullable: true
          description: 終えたカード。無ければ null
        started_drip_id:
          type: string
          format: uuid
          nullable: true
          description: 始めたカード。待機が無ければ null
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid order ID format"
...